fix(subscription/reload): calculate total hours manually
The backend expects the duration of a subscription package in the
format `HH:mm:ss`. Our  previous implementation failed if the
duration exceeded 24 hours.

For instance, if the duration was 30 hours, it was given in the
format `DD HH:mm:ss` which the backend did not expect. Now, the
calculation of total hours is done manually and not via Moment.js,
as there is not a feasible solution for `DD HH:mm:ss` -> `HH:mm:ss`.

Furthermore, the transform has been updated to not throw an error
when serializing a non-moment-duration value.